pyproject.toml
[project]
name = "xgboost-poc"
version = "0.1.0"
description = "Engaging PoC explaining the workings of XGBoost via script-based demos"
authors = [{ name = "Master Lonn" }]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

# Core runtime dependencies (kept light to avoid build conflicts on Py3.11)
dependencies = [
  "xgboost>=2.0",
  "numpy>=1.26,<2.0",
  "pandas>=2.0",
  "scikit-learn>=1.3,<1.6",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
# Explainability extras (installed on-demand): SHAP + compatible JIT stack
explain = [
  "shap>=0.46",
  "numba>=0.58",
  "llvmlite>=0.41",
]

# Developer tooling and optional interactive extras
dev = [
  "jupyterlab>=4.0",
  "ipykernel",
  "ruff>=0.5",
  "black>=24.0",
  "mypy>=1.7",
  "pytest>=7.4",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
